Shorts Model (Table) This model represents individual short content items and includes:
A primary key id, again generated with cuid() for uniqueness.

Fields for title, description, and url, which hold metadata about the short.

userId is a foreign key linking each short to its creator.

createdAt and updateAt automatically track when the short is made or changed.

A user relation that maps the userId to the id of the User model, establishing a foreign key relationship.

-> Relations
There’s a bidirectional one-to-many relationship:

On the User side, the shorts field is an array of Shorts items.

On the Shorts side, the userId links back to the User, with a foreign key constraint.

Prisma ensures referential integrity between the two tables.

-> Timestamps
Both models track creation and updates with createdAt (set on creation) and updatedAt (updated on every modification).

-> Unique Constraints and Defaults
clerkUserId is uniquely enforced to prevent duplicate users from being linked to the same external identity.

Defaults like now() and cuid() ensure that IDs and timestamps are automatically handled by Prisma.